Move litaestetik modules and directories to aestetik
